renamed:    pyrpod/plume/PlumeStrikeEstimationStudy.py -> pyrpod/plume/PlumeStrikeCalculator.py
