+++ pyrpod/plume/PlumeStrikeCalculator.py
