+++ pyrpod/rpod/PlumeStrikeEstimationStudy.py
@@ -27,7 +27,7 @@ from pyrpod.rpod.approach_maneuvers import (
 )
 from pyrpod.rpod.io import ensure_results_dirs, write_jfh
 from pyrpod.rpod.PlumeStudyExport import PlumeStudyExport
-from pyrpod.plume.PlumeStrikeEstimationStudy import compute_plume_strikes
+from pyrpod.plume.PlumeStrikeCalculator import compute_plume_strikes
 
 logger = get_logger("pyrpod.rpod.PlumeStrikeEstimationStudy")
